pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "eagle-trading"
version = "1.0.0"
description = "Eagle: A highly versatile, adaptable, trainable, efficient, and intelligent AI-powered trading bot"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Eagle Team"}
]
keywords = ["trading", "bot", "ai", "machine-learning", "quantitative-finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "torch>=2.1.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "flask>=3.0.0",
    "ta>=0.11.0",
    "yfinance>=0.2.30",
    "plotly>=5.18.0",
    "duckdb>=0.9.0",
    "python-dotenv>=1.0.0",
    "ccxt>=4.1.0",
    "alpaca-py>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
eagle = "eagle.cli:main"

[tool.setuptools.packages.find]
include = ["eagle*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
